+++ app/src/main/java/com/example/simpsons/features/characters/data/CharactersDataRepository.kt
@@ -1,7 +1,7 @@
 package com.example.simpsons.features.characters.data
 
 import com.example.simpsons.features.characters.data.remote.api.CharactersApiRemoteDataSource
-import com.example.simpsons.features.characters.domain.Character
+import com.example.simpsons.features.characters.domain.model.Character
 import com.example.simpsons.features.characters.domain.CharactersRepository
 
 class CharactersDataRepository(private val charactersApiRemoteDataSource: CharactersApiRemoteDataSource): CharactersRepository {
