+++ app/src/main/java/com/example/simpsons/features/characters/data/remote/api/CharactersApiService.kt
@@ -1,5 +1,6 @@
 package com.example.simpsons.features.characters.data.remote.api
 
+import com.example.simpsons.features.characters.data.remote.api.model.ApiResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
