+++ app/src/main/java/com/example/simpsons/features/characters/domain/CharactersRepository.kt
@@ -1,5 +1,7 @@
 package com.example.simpsons.features.characters.domain
 
+import com.example.simpsons.features.characters.domain.model.Character
+
 interface CharactersRepository {
     suspend fun getAllCharactersList(pageNumber: Int): Result<List<Character>>
 }
