+++ app/src/main/res/layout/fragment_simpsons_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".features.character_details.presentation.SimpsonsDetailsFragment">
+    tools:context=".features.characters.presentation.details.SimpsonsDetailsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
