+++ app/src/main/res/layout/fragment_simpsons_list.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:theme="@style/AppTheme"
-    tools:context=".features.characters.presentation.SimpsonsListFragment">
+    tools:context=".features.characters.presentation.list.SimpsonsListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fslRvCharactersList"
