+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/fragment_list"
-        android:name="com.example.simpsons.features.characters.presentation.SimpsonsListFragment"
+        android:name="com.example.simpsons.features.characters.presentation.list.SimpsonsListFragment"
         android:label="@string/list_fragment"
         tools:layout="@layout/fragment_simpsons_list">
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/fragment_details"
-        android:name="com.example.simpsons.features.character_details.presentation.SimpsonsDetailsFragment"
+        android:name="com.example.simpsons.features.characters.presentation.details.SimpsonsDetailsFragment"
         android:label="@string/details_fragment"
         tools:layout="@layout/fragment_simpsons_details">
         <argument
